refactor(notification): mv adapter pkg
